Project Adjustments (MENU) pt.2
On the previous Commit I said that this would be a 2 part push. However things have changed a bit and I'm now just gonna do it in 3 parts.

- All menu functions returned. Inject and have fun
- GUI Launcher is broken, please use a previous build until i push the next commit which will be focused on the GUI Launcher.

The next and final push is focused on the Launcher side of things. I will be pointing all the libs to the root lib folder as well as making adjustments to the launcher not closing on launch. I also plan to include a method to either keep the launcher open or allow the user to inject into the process multiple times now that a method to unload the dll has been included.