Project Adjustments (#9)
* Project Adjustments (MENU) pt.1 ----------
This is a 2 part commit. The first piece will include adjustments revolving around the menu and adjustments to the main repository page

- Better Repository Organization
- Retain 1 Click Build
- Adjusted build output to "bin" folder
- Combined Projects

* Project Adjustments (MENU) pt.2 ----------
On the previous Commit I said that this would be a 2 part push. However things have changed a bit and I'm now just gonna do it in 3 parts.

- All menu functions returned. Inject and have fun
- GUI Launcher is broken, please use a previous build until i push the next commit which will be focused on the GUI Launcher.

The next and final push is focused on the Launcher side of things. I will be pointing all the libs to the root lib folder as well as making adjustments to the launcher not closing on launch. I also plan to include a method to either keep the launcher open or allow the user to inject into the process multiple times now that a method to unload the dll has been included.

* Project Updates (LAUNCHER) pt.3 ----------
This commit revolves around bringing the GUI Launcher File Directory up to date with the new repository layout.

- 1 Click Build restored

People can now open the project , right click the solution "EROverlay" and select build. It will build both the launcher and the DLL and output to "bin"

* Project Adjustments (LAUNCHER) pt.4 ----------
Final adjustments made. Ready for merge
- Launcher now exits on successful process launch